Major update to large_page library and example
Now large_page supports MapDSOToLargePages which takes a const char*
argument which is regex-parsed and used to identify .so libraries
and map them to large pages.
Minor code cleanup and refactoring to support the new functionality.
Changed ld.implicit.script and add alignment to .text and .lpstub.
Changed example which now supports optional command line arguments:
large_page_example [soname1] [soname2] [...]
For example:
./large_page_example libc ld
will try to map libc.so and ld.so to large pages in addition to the
static code.