source: Move map_status strings to library
Add MapStatusStr which returns a string representing the map_status
value given as parameter. The second parameter controls whether it
returns a short version (the name of the error as it appears in the
map_status enum), or the actual description of the error.
Add two macros to simplify usage: MAP_STATUS_STR(status) will return
the full description, MAP_STATUS_STR_SHORT(status) will return
the shorter version.
Move map_status_text from large_page_example.c to large_page.c and
expand it to include the short versions of the strings.