large pages: define reference symbol via asm
We replace the linker script, which is not portable, with an assembly
file that defines a reference symbol to be placed at the beginning of
the .text section. We rely on the fact that the linker tends to place
symbols in the order in which it encounters them and on the fact that
the object file resulting from the assembly file is the first one the
linker encounters to ensure that the symbol is placed early in the text
section.

Signed-off-by: Gabriel Schulhof <gabriel.schulhof@intel.com>
PR-URL: https://github.com/intel/iodlr/pull/25
Reviewed-By: Suresh Srinivas <suresh.srinivas@intel.com>