tools: remove numbered index from help metrics list
The help prints out the list of available metrics as number
indexed list:

```
  -m metrics : Use comma separator to specify one or many of the
             following metrics
    0) itlb_stalls [Default]
    1) icache_miss_stalls
    2) itlb_mpki
    3) l1_code_read_MPI
    4) l2_demand_code_MPI
```

which misled me to believe you could, or should, use the index number on
the command line to specify the metrics - which you cannot - you must
use the 'name'.

Drop the number indexing, as afaict it is not used anywhere else, and only
adds potential confusion to the commandline help. Output now looks like:

```
  -m metrics : Use comma separator to specify one or many of the
             following metrics
    - itlb_stalls [Default]
    - icache_miss_stalls
    - itlb_mpki
    - l1_code_read_MPI
    - l2_demand_code_MPI
```

Signed-off-by: Graham Whaley <graham.whaley@intel.com>