tools: enable pid and command together
It can sometimes be useful to run a command (benchmark say), and
monitor a differnet PID (server say). Currently the
measure-perf-metric.sh command only allows one or the other, and
if both are supplied, silintly drops the first.

Enable use of both -p (PID) and -e (command) together, as this is
a valid perf mode.

Signed-off-by: Graham Whaley <graham.whaley@intel.com>